Modularize init command
