+++ src/commands/init/getShortDnpName.ts
@@ -1,0 +1,6 @@
+import { validateDnpName } from "./validateDnpName";
+
+export function getShortDnpName(dnpName: string): string {
+    validateDnpName(dnpName);
+    return dnpName.split(".")[0];
+}
