+++ src/commands/init/index.ts
@@ -1,0 +1,1 @@
+export { init } from "./init.js";
